requirements.txt
aiofiles==23.2.1
altair==5.3.0
amqp==5.1.1
anaconda==0.0.1.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
bleach==6.0.0
cachetools==5.3.0
celery==5.2.7
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colab==1.13.5
comm==0.2.0
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diazo==1.5.0
Django==1.7.11
django-haystack==2.4.1
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==0.22.2
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
etiquetando==0.1
exceptiongroup==1.2.2
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.3
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
google-api-core==1.34.0
google-auth==2.16.2
google-cloud-pubsub==2.1.0
googleapis-common-protos==1.58.0
gradio==4.38.1
gradio_client==1.1.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.48.2
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_resources==6.4.0
ipykernel==6.26.0
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.4
Js2Py==0.74
json5==0.9.14
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.2.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kombu==5.2.4
libcst==0.4.9
loguru==0.7.2
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.6
notebook_shim==0.2.3
numpy==2.0.0
opencv-contrib-python==4.7.0.72
orjson==3.10.6
overrides==7.4.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pip-upgrade-tool==0.7.4
platformdirs==3.1.1
portpicker==1.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyjsparser==2.7.1
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.0.1
redis==4.5.1
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.5.1
scipy==1.9.3
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.4
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.37.2
stemming==1.0.1
terminado==0.13.3
tinycss2==1.2.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.9.0
typer==0.12.3
types-python-dateutil==2.8.19.14
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vine==5.0.0
watchfiles==0.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Whoosh==2.7.4
